Issue 51118 - UI - improve modal validation when creating an instance
Description:  Do not enable the "create" button until all the fields are
              valid (DN's, port numbers, passwords, etc).

              Improve layout and handling of optional database settings.

              Add a json argument to dscreate so the UI can report any
              failure text.  Also improve error reporting in dscreate.

relates: https://pagure.io/389-ds-base/issue/51118

Reviewed by: firstyear & spichugi(Thanks!!)

Improve validation error messages

Fix allowed characters
